cisco_ise_networkdevcie.py - upgrade
The proposed changes are:
-	Introducing new ansible variable 
delete_devices: true/false
with true – all devices listed will be removed (as your current code looks like to work in a way that by using “cisco_ise_networkdevice_force: true” you were deleting and re adding it)
-	Update function – changed
when “delete_devices” is false we will try to add new or if existing- update device. 
Update will be done only if  1 device exists on ISE. So based on IP and device name provided by user, code is checking for both. 
There are 3 options basically:
1)	No devices exist 
2)	1 device exists
3)	2 devices exist (1s for the same name and 2nd for the same IP)
Update function will be invoked only if 1 device exists and it is done by merging what user entered in the play and what is already defined on ISE (your code basically was deleting existing one and re-adding it which in result could cause the loss of information already defined on ISE for a specific device) 
-	Now we are able to add mask to the device definition so larger objects can be added (subnets) – if not provided, default /32 is taken the same way you have in your code. 
-	We moved away from xml into json style – as it was easier for us + that also eliminated the same problem other guys reported to you on endpoint tasks that he had to change resource with {ers.ise.cisco.com}resource As getting and empty result. (we noticed the same thing)
{ers.ise.cisco.com}resource
-	There is not more need for that part as described above we are not re-adding it

# if true all devices will be deleted and created again
# use this if only RADIUS or TACACS shared secrets changed
cisco_ise_networkdevice_force: false